fixed continual adding of delimiter

+++ week09hw_v3_check.py
@@ -98,15 +98,17 @@ else:
         edit_file = open(f"{directory}/{filename}", "r+")
         file_contents = edit_file.read()
         # Make changes to file_contents as needed
-        file_contents = file_contents.replace("SET", "-- ~\nSET")
-        file_contents = file_contents.replace("USE", "-- ~\nUSE")
-        file_contents = file_contents.replace("SELECT", "-- ~\nSELECT")
-        file_contents = file_contents.replace(";", ";\n-- ~")
-        edit_file.seek(0)
-        edit_file.write(file_contents)
-        edit_file.truncate()
-        edit_file.close()
-        
+        if not file_contents.__contains__('-- ~'):
+            file_contents = file_contents.replace("^USE", "-- ~\nUSE")
+            file_contents = file_contents.replace("^SELECT", "-- ~\nSELECT")
+            file_contents = file_contents.replace(";$", ";\n-- ~")
+            edit_file.seek(0)
+            edit_file.write(file_contents)
+            edit_file.truncate()
+            edit_file.close()
+        else:
+            edit_file.close()
+
         f = open(f"{directory}/{filename}", "r")
             
         answer.write("***********************************\n")
