+++ week10hw_v3_check.py
@@ -120,15 +120,19 @@ else:
         file_count += 1 # increment the counter
         edit_file = open(f"{directory}/{filename}", "r+")
         file_contents = edit_file.read()
-        # Make changes to file_contents as needed
-        file_contents = file_contents.replace("SET", "-- ~\nSET")
-        file_contents = file_contents.replace("USE", "-- ~\nUSE")
-        file_contents = file_contents.replace("SELECT", "-- ~\nSELECT")
-        file_contents = file_contents.replace(";", ";\n-- ~")
-        edit_file.seek(0)
-        edit_file.write(file_contents)
-        edit_file.truncate()
-        edit_file.close()
+        # check to see if delimiter exists
+        if not file_contents.__contains__('-- ~'):
+            file_contents = file_contents.replace("SET", "-- ~\nSET")
+            file_contents = file_contents.replace("USE", "-- ~\nUSE")
+            file_contents = file_contents.replace("SELECT", "-- ~\nSELECT")
+            file_contents = file_contents.replace(";", ";\n-- ~")
+            edit_file.seek(0)
+            edit_file.write(file_contents)
+            edit_file.truncate()
+            edit_file.close()
+        else:
+            # print("Delimiter Found")
+            pass
 
         f = open(f"{directory}/{filename}", "r")
